+++ NEWS.md
@@ -1,6 +1,9 @@
 ## CHANGES IN udpipe VERSION 0.8.2
 
 - document_term_matrix.data.frame now preserves order of the documents (issue #44)
+- dtm_remove_lowfreq, dtm_remove_tfidf, dtm_remove_terms gain extra argument remove_emptydocs
+  explicitely add drop=FALSE to internal dtm_... calls
+- add dtm_remove_sparseterms (issue #44)
 
 ## CHANGES IN udpipe VERSION 0.8.1